# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Akmatori API
  description: |
    AI-powered AIOps platform API. Manages skills, tools, incidents, alert sources,
    and system settings for automated incident investigation and remediation.
  version: 1.0.0
  contact:
    name: Akmatori
servers:
  - url: /api
    description: API base path
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
        details:
          type: object
          additionalProperties:
            type: string
          description: Field-level validation errors
...
    Skill:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          description: Kebab-case skill name
        description:
          type: string
        category:
          type: string
        is_system:
          type: boolean
        enabled:
          type: boolean
        prompt:
          type: string
          description: SKILL.md prompt content
        tools:
          type: array
          items:
            $ref: '#/components/schemas/ToolInstance'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ToolType:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        schema:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ToolInstance:
      type: object
      properties:
        id:
          type: integer
        tool_type_id:
          type: integer
        name:
          type: string
        settings:
          type: object
        enabled:
          type: boolean
        tool_type:
          $ref: '#/components/schemas/ToolType'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /skills:
    get:
      summary: List skills
      operationId: listSkills
      tags: [Skills]
      responses:
        '200':
          description: List of skills with prompts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Skill'
    post:
      summary: Create skill
      operationId: createSkill
      tags: [Skills]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 64
                description:
                  type: string
                  maxLength: 1024
                category:
                  type: string
                  maxLength: 64
                prompt:
                  type: string
      responses:
        '201':
          description: Skill created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
        '400':
          $ref: '#/components/responses/BadRequest'
  /skills/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get skill
      operationId: getSkill
      tags: [Skills]
      responses:
        '200':
          description: Skill with prompt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update skill
      operationId: updateSkill
      tags: [Skills]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                category:
                  type: string
                enabled:
                  type: boolean
                prompt:
                  type: string
      responses:
        '200':
          description: Updated skill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete skill
      operationId: deleteSkill
      tags: [Skills]
      responses:
        '204':
          description: Skill deleted
        '403':
          description: Cannot delete system skill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /skills/{name}/prompt:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get skill prompt
      operationId: getSkillPrompt
      tags: [Skills]
      responses:
        '200':
          description: Prompt content
          content:
            application/json:
              schema:
                type: object
                properties:
                  prompt:
                    type: string
    put:
      summary: Update skill prompt
      operationId: updateSkillPrompt
      tags: [Skills]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                prompt:
                  type: string
      responses:
        '200':
          description: Prompt updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /skills/{name}/tools:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get skill tools
      operationId: getSkillTools
      tags: [Skills]
      responses:
        '200':
          description: Tool instances assigned to skill
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolInstance'
    put:
      summary: Update skill tools
      operationId: updateSkillTools
      tags: [Skills]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                tool_instance_ids:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          description: Updated skill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
...
  /tools:
    get:
      summary: List tool instances
      operationId: listTools
      tags: [Tools]
      responses:
        '200':
          description: Tool instances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolInstance'
    post:
      summary: Create tool instance
      operationId: createTool
      tags: [Tools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tool_type_id, name]
              properties:
                tool_type_id:
                  type: integer
                name:
                  type: string
                settings:
                  type: object
      responses:
        '201':
          description: Tool created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolInstance'
  /tools/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Get tool instance
      operationId: getTool
      tags: [Tools]
      responses:
        '200':
          description: Tool instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolInstance'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update tool instance
      operationId: updateTool
      tags: [Tools]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                settings:
                  type: object
                enabled:
                  type: boolean
      responses:
        '200':
          description: Updated tool
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolInstance'
    delete:
      summary: Delete tool instance
      operationId: deleteTool
      tags: [Tools]
      responses:
        '204':
          description: Tool deleted
...